Validate URLs in web commands before execution (#2616)
